requirements.txt
django
django-q
gunicorn
djangorestframework
